feat: enhance theme toggle functionality and improve UI components
